added 3 test cases and supporting classes
