+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/projects/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/jaxrs/RequestBean.java
@@ -1,0 +1,14 @@
+package io.openliberty.sample.jakarta.jaxrs;
+
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.PathParam;
+
+public class RequestBean {
+    @PathParam("userId")
+    private String userId;
+
+    @QueryParam("sort")
+    private String sort;
+
+}
+
