+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/projects/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/jaxrs/SessionBean.java
@@ -1,0 +1,13 @@
+package io.openliberty.sample.jakarta.jaxrs;
+
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.PathParam;
+
+public class SessionBean {
+    @PathParam("sessionId")
+    private String sessionId;
+
+    @QueryParam("invalidateSession")
+    private boolean invalidateSession;
+
+}
